+++ ChangeLog.md
@@ -21,6 +21,7 @@
 * !cof-effet-temp met bien l'état associé quand on utilise paralyseTemp ou etourdiTemp
 
 ### Autres améliorations
+* Affichage du nombre de soins restants dans le statut
 * ajout d'explications aux jets de caractéristiques
 * ajout d'une option --secret à !cof-attack